refactor: Reorganize order-related files structure
