+++ order-service/src/main/java/com/ed/orderservice/application/service/domain/order/create/OrderPaymentCreateService.java
@@ -1,11 +1,10 @@
-package com.ed.orderservice.application.service;
+package com.ed.orderservice.application.service.domain.order.create;
 
 
 import com.ed.OrderPaymentCreateRequestEvent;
 import com.ed.orderservice.application.port.out.OrderPaymentCreateOutPort;
 import com.ed.orderservice.application.port.out.OrderStatusUpdateOutPort;
-import com.ed.orderservice.application.service.serializer.OrderPaymentRequestSerializer;
-import com.ed.orderservice.domain.vo.order.Order;
+import com.ed.orderservice.application.service.domain.order.create.serializer.OrderPaymentRequestSerializer;
 import com.ed.orderservice.domain.vo.order.OrderEvent;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -20,8 +19,8 @@ public class OrderPaymentCreateService {
   private final OrderStatusUpdateOutPort  orderStatusUpdateOutPort;
   private final OrderPaymentRequestSerializer orderPaymentRequestSerializer;
 
-  public void requestPaymentConfirmation(OrderEvent orderEvent, Order order) {
-    orderStatusUpdateOutPort.updateToPaymentRequest(order.getOrderPublicId());
+  public void requestPaymentConfirmation(OrderEvent orderEvent, String orderId) {
+    orderStatusUpdateOutPort.updateToPaymentRequest(orderId);
     orderPaymentCreateOutPort.sendPaymentConfirmRequest(deserializeOrderPaymentRequest(orderEvent));
   }
