+++ order-service/src/main/java/com/ed/orderservice/application/service/domain/order/create/serializer/OrderPaymentRequestSerializer.java
@@ -1,4 +1,4 @@
-package com.ed.orderservice.application.service.serializer;
+package com.ed.orderservice.application.service.domain.order.create.serializer;
 
 import com.ed.OrderPaymentCreateRequestEvent;
 import com.ed.orderservice.domain.vo.order.Order;
