+++ order-service/src/main/java/com/ed/orderservice/presentaion/web/response/domain/create/NewPurchaseOrderMessageResponse.java
@@ -1,4 +1,4 @@
-package com.ed.orderservice.presentaion.web.response.domain.orderCreate;
+package com.ed.orderservice.presentaion.web.response.domain.create;
 
 import com.ed.orderservice.application.port.in.dto.OrderItemDto;
 import com.ed.orderservice.domain.vo.order.Order;
