+++ order-service/src/main/java/com/ed/orderservice/presentaion/web/response/domain/settlement/OrderSettlementResponse.java
@@ -1,4 +1,4 @@
-package com.ed.orderservice.presentaion.web.response.domain.orderSettlement;
+package com.ed.orderservice.presentaion.web.response.domain.settlement;
 
 
 import static lombok.AccessLevel.PRIVATE;
